assignment-7/requirements.txt
Flask==1.1.2
google-cloud-datastore==1.12.0
google-auth
requests_oauthlib